% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gen.R
\name{add_gen}
\alias{add_gen}
\title{Add Generalization Relationship to Hierarchy}
\usage{
add_gen(from, to, hierachy)
}
\arguments{
\item{from}{A character or numeric value indicating the node to be generalized.}

\item{to}{A character or numeric value indicating the parent node to which the \code{from} node will be generalized.}

\item{hierachy}{A list representing the current hierarchy structure.}
}
\value{
A modified hierarchy list with the new generalization relationship added.
}
\description{
This function adds a generalization relationship to a given hierarchy.
It assigns a parent node to a specified node in the hierarchy.
}
\examples{
# Example usage:
hierarchy <- list()
hierarchy <- add_gen("child", "parent", hierarchy)
print(hierarchy)
}
\keyword{internal}
