% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_min_size_of_equivalence_class.r
\name{compute_min_size_of_equivalence_class}
\alias{compute_min_size_of_equivalence_class}
\title{Compute the Minimum Set Size of an Equivalence Class}
\usage{
compute_min_size_of_equivalence_class(data, quid)
}
\arguments{
\item{data}{A data frame or data table containing the quasi-identifiers}

\item{quid}{A vector of column names or indices representing quasi-identifiers}
}
\value{
The minimum frequency (k) of any combination of quasi-identifier values
}
\description{
This function calculates the minimum frequency records that belong to the same constellation of quasi-identifiers (i.e. equivalence class).
It is useful for evaluating k-anonymity, where k is the minimum size of any equivalence class.
}
\examples{
# For a data frame with columns "age", "gender", "zipcode"
# To compute min size where "age" and "zipcode" are quasi-identifiers:
# compute_min_size_of_equivalence_class(my_data, c("age", "zipcode"))

}
