% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_childrens.R
\name{find_childrens}
\alias{find_childrens}
\title{Find Children of a Node in a Hierarchy}
\usage{
find_childrens(node, hierarchy)
}
\arguments{
\item{node}{}
}
\description{
Find Children of a Node in a Hierarchy
}
\keyword{internal}
