% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_parent.R
\name{find_parent}
\alias{find_parent}
\title{Finds the parent node of a given node in a hierarchy.}
\usage{
find_parent(node, hierarchy)
}
\arguments{
\item{node}{A character the name of the node for which the parent node is to be found.}

\item{hierarchy}{A list representing the current hierarchy structure.}
}
\value{
The parent node of the given node in the hierarchy.
}
\description{
Finds the parent node of a given node in a hierarchy.
}
\keyword{internal}
