% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize_child.R
\name{generalize_child}
\alias{generalize_child}
\title{Generalize a child node to its parent node in a given attribute}
\usage{
generalize_child(attribute, node, Parent)
}
\arguments{
\item{attribute}{A factor attribute}

\item{node}{A child node to be generalized}

\item{Parent}{A parent node to generalize the child node to}
}
\value{
A factor attribute with the child node generalized to the parent node
}
\description{
Generalize a child node to its parent node in a given attribute
}
\keyword{internal}
