% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hierachy.r
\name{plot_hierarchy}
\alias{plot_hierarchy}
\title{Plot Hierarchical Structure}
\usage{
plot_hierarchy(
  hierarchy_list,
  var,
  vertex_size = 10,
  edge_arrow_size = 0.5,
  main = NULL,
  ...
)
}
\arguments{
\item{hierarchy_list}{A list containing hierarchical relationships.}

\item{var}{The variable name (i.e. single quasi-identifier) within the hierarchy list to plot.}

\item{vertex_size}{Size of the vertices. Default is 10.  (optional)}

\item{edge_arrow_size}{Size of the edge arrows. Default is 0.5. (optional)}

\item{main}{Title of the plot. If not given is generated. (optional)}
}
\value{
An igraph object representing the hierarchy.
}
\description{
Visualize the hierachical structure used to generalize quasi-identifiers.
}
\examples{
plot_hierarchy(taxonomy_adult, "age")

}
