% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse_hierachy.R
\name{traverse_hierachy}
\alias{traverse_hierachy}
\title{Traverse hierachy}
\usage{
traverse_hierachy(children, attribute, hierarchy)
}
\arguments{
\item{children}{}

\item{attribute}{}

\item{hierarchy}{A list representing the current hierarchy structure.}
}
\value{
A modified attribute list with the new generalization relationship added.
}
\description{
This is an internal function that generalizes all children of a parent node according to a specified hierarchy.
}
\keyword{internal}
